refactor(charging-station): extract config/template helpers to HelpersConfig (#1936 f-5)
Extracts 10 configuration/template helpers (buildTemplateName, validateStationInfo,

getDefaultConnectorMaximumPower, checkConfiguration, setChargingStationOptions,

stationTemplateToStationInfo, getAmperageLimitationUnitDivider, getDefaultVoltageOut,

getIdTagsFile, getEvProfilesFile) plus the two template-topology helpers

getMaxNumberOfEvses and getMaxNumberOfConnectors into HelpersConfig.ts.

The two template-topology helpers were originally listed as core-station helpers

in the issue body, but getDefaultConnectorMaximumPower consumes both of them;

moving them into HelpersConfig avoids a cross-module dependency back into

Helpers.ts that would risk an ESM live-binding cycle. Helpers.ts imports

getMaxNumberOfConnectors from HelpersConfig for use in the retained

getConfiguredMaxNumberOfConnectors, closing the split with no cycle.

Helpers.ts is now 185 LOC (core station helpers + 5 barrel re-exports),

down from 1389 LOC before the phased split.

Refs #1936
